requirements.txt
gradio
python-dotenv
azure-cognitiveservices-speech
fastapi
uvicorn
sqlalchemy
itsdangerous
passlib
bcrypt